requirements_production.txt
# BabyShield Production Requirements
# Python 3.11+

# === CORE FRAMEWORK ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]

# === DATABASE ===
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# === CACHING ===
redis==6.4.0
hiredis==2.3.2

# === ASYNC PROCESSING ===
celery==5.3.4
celery[sqs]  # For AWS SQS
kombu==5.3.4

# === API ESSENTIALS ===
python-multipart==0.0.9
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# === SECURITY ===
slowapi==0.1.9  # Rate limiting
python-jose[cryptography]==3.3.0
cryptography==41.0.7
bcrypt==4.1.2

# === AWS INTEGRATION ===
boto3==1.34.0
aioboto3==12.3.0
botocore==1.34.0

# === IMAGE PROCESSING ===
Pillow==11.3.0
opencv-python-headless==4.9.0.80  # Use headless for production
pyzbar==0.1.9
pylibdmtx==0.1.10
qrcode[pil]==7.4.2
python-barcode==0.15.1

# === OCR & VISION ===
google-cloud-vision==3.5.0
pytesseract==0.3.10
easyocr==1.7.1

# === MONITORING & LOGGING ===
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.0
datadog==0.47.0
structlog==24.1.0
python-json-logger==2.0.7

# === PERFORMANCE ===
orjson==3.9.10
ujson==5.9.0
uvloop==0.19.0
httptools==0.6.1
aiodns==3.1.1
cchardet==2.1.7

# === RESILIENCE ===
pybreaker==1.0.1
tenacity==8.2.3
backoff==2.2.1

# === HTTP CLIENTS ===
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# === UTILITIES ===
python-dotenv==1.0.0
python-dateutil==2.9.0
pytz==2023.3
pyyaml==6.0.1

# === TESTING (for load testing) ===
locust==2.20.0
pytest==7.4.4
pytest-asyncio==0.21.1

# === PRODUCTION SERVER ===
gunicorn==21.2.0
gevent==23.9.1
eventlet==0.33.3

# === DATA PROCESSING ===
pandas==2.1.4
numpy==1.26.3

# === DOCUMENTATION ===
reportlab==4.0.9
weasyprint==60.2
xhtml2pdf==0.2.11

# === CONNECTORS ===
feedparser==6.0.11  # For RSS feeds
beautifulsoup4==4.12.2
lxml==5.0.0

# === SERIALIZATION ===
msgpack==1.0.7
marshmallow==3.20.1

# === CORS & MIDDLEWARE ===
fastapi-cors==0.0.6
starlette-context==0.3.6
asgi-correlation-id==4.3.0

# === DEPLOYMENT ===
python-multipart==0.0.9
websockets==12.0
watchfiles==0.21.0

# === OPTIONAL OPTIMIZATIONS ===
# numba==0.58.1  # For numerical optimizations
# cython==3.0.7  # For C-level optimizations
